Attempted to fix tooltip copy/paste issue.
In the document contents view, users would like to be able to copy information from the tooltips that appear when you hover over a word. In the previous version of the application, this could be done by clicking on a word, which caused the tooltip to stay visible. This was a first attempt at reintroducing that behaviour, unfortunately unsuccesful.